--- airflow_dags/images/airflow-ml-base/Dockerfile
@@ -1,7 +1,0 @@
-FROM python:3.6-slim
-
-RUN mkdir /build/
-
-COPY requirements.txt /build/requirements.txt
-
-RUN pip install -r /build/requirements.txt
