--- airflow_dags/images/airflow-ml-base/requirements.txt
@@ -1,6 +1,0 @@
-click
-mlflow==1.17.0
-numpy
-pandas
-psycopg2-binary==2.8.6
-scikit-learn==0.24.1
