--- airflow_dags/images/airflow-split/split.py
@@ -1,23 +1,0 @@
-import os
-
-import click
-import pandas as pd
-from sklearn.model_selection import train_test_split
-
-
-@click.command("split")
-@click.option("--input-dir")
-@click.option("--output-dir")
-def split(input_dir: str, output_dir: str):
-    data = pd.read_csv(os.path.join(input_dir, "data.csv"), index_col=0)
-    target = pd.read_csv(os.path.join(input_dir, "target.csv"), index_col=0)
-    data = data.merge(target, left_index=True, right_index=True)
-    train, val = train_test_split(data, test_size=0.3)
-    
-    os.makedirs(output_dir, exist_ok=True)
-    train.to_csv(os.path.join(output_dir, "train.csv"))
-    val.to_csv(os.path.join(output_dir, "val.csv"))
-    
-
-if __name__ == "__main__":
-    split()
