--- airflow_dags/images/mlflow/Dockerfile
@@ -1,5 +1,0 @@
-FROM python:3.6-slim
-
-COPY requirements.txt /home/requirements.txt
-
-RUN pip install -r /home/requirements.txt --no-cache-dir
