--- airflow_dags/images/mlflow/requirements.txt
@@ -1,2 +1,0 @@
-mlflow==1.17
-psycopg2-binary==2.8
