+++ airflow_ml_dags/images/airflow-download/download.py
@@ -1,0 +1,17 @@
+import os
+
+import click
+from sklearn.datasets import load_breast_cancer
+
+
+@click.command("download")
+@click.argument("output_dir")
+def download(output_dir: str):
+    X, y = load_breast_cancer(return_X_y=True, as_frame=True)
+    os.makedirs(output_dir, exist_ok=True)
+    X.to_csv(os.path.join(output_dir, "data.csv"))
+    y.to_csv(os.path.join(output_dir, "target.csv"))
+
+
+if __name__ == "__main__":
+    download()
