+++ airflow_ml_dags/images/airflow-ml-base/Dockerfile
@@ -1,0 +1,5 @@
+FROM python:3.6-slim
+RUN mkdir /build/
+COPY requirements.txt /build/requirements.txt
+
+RUN pip install -r /build/requirements.txt --no-cache-dir
