+++ airflow_ml_dags/images/airflow-ml-base/requirements.txt
@@ -1,0 +1,6 @@
+scikit-learn==0.24.1
+click
+pandas
+numpy
+mlflow==1.17.0
+psycopg2-binary==2.8.6
