+++ airflow_ml_dags/images/airflow-preprocess/Dockerfile
