+++ airflow_ml_dags/images/mlflow/Dockerfile
@@ -1,0 +1,4 @@
+FROM python:3.6-slim-stretch
+COPY requirements.txt /home/requirements.txt
+
+RUN pip install -r /home/requirements.txt --no-cache-dir
