+++ airflow_ml_dags/images/mlflow/requirements.txt
@@ -1,0 +1,2 @@
+mlflow~=1.17
+psycopg2-binary~=2.8
